+++ tests/test_blob.py
@@ -71,6 +71,19 @@ class WordListTest(TestCase):
         wl = tb.WordList(self.words)
         assert_equal(list(wl), self.words)
 
+    def test_append(self):
+        wl = tb.WordList(['dog'])
+        wl.append("cat")
+        assert_true(isinstance(wl[1], tb.Word))
+        wl.append(('a', 'tuple'))
+        assert_true(isinstance(wl[2], tuple))
+
+    def test_extend(self):
+        wl = tb.WordList(["cats", "dogs"])
+        wl.extend(["buffalo", 4])
+        assert_true(isinstance(wl[2], tb.Word))
+        assert_true(isinstance(wl[3], int))
+
 
 class SentenceTest(TestCase):
 
@@ -81,7 +94,7 @@ class SentenceTest(TestCase):
 
     def test_repr(self):
         assert_equal(repr(self.sentence),
-                     "Sentence('{0}')".format(self.raw_sentence))
+                     "Sentence({0})".format(repr(self.raw_sentence)))
 
     def test_stripped_sentence(self):
         assert_equal(self.sentence.stripped,
@@ -263,11 +276,7 @@ is managed by the non-profit Python Software Foundation.'''
 
     def test_repr(self):
         blob1 = tb.TextBlob('lorem ipsum')
-        assert_equal(repr(blob1), "TextBlob('lorem ipsum')")
-        big_blob = tb.TextBlob(self.text)
-        assert_equal(repr(big_blob),
-            "TextBlob('Beautiful is better than ugly.\nExplicit is better ...'s do more of those!')"
-        )
+        assert_equal(repr(blob1), "TextBlob({0})".format(repr('lorem ipsum')))
 
     def test_cmp(self):
         blob1 = tb.TextBlob('lorem ipsum')
@@ -597,7 +606,8 @@ is managed by the non-profit Python Software Foundation.'''
     @attr('slow')
     def test_json(self):
         blob = tb.TextBlob('Beautiful is better than ugly. ')
-        blob_dict = json.loads(blob.json())[0]
+        assert_equal(blob.json, blob.to_json())
+        blob_dict = json.loads(blob.json)[0]
         assert_equal(blob_dict['stripped'], 'beautiful is better than ugly')
         assert_equal(blob_dict['noun_phrases'], blob.sentences[0].noun_phrases)
         assert_equal(blob_dict['start_index'], blob.sentences[0].start)